requirements.txt
fastapi
uvicorn
redis
httpx
pydantic
pytest
pytest-asyncio
anyio
google-cloud-pubsub
